Allows all instances in VPC to communicate
The firewall rule was previously incorrect, only allowing instanceds in
the same 10.0.0.0/16 prefix to communicate, which would only let
instances in a single subnet communicate. Our subnets are 10.0.0.0/16,
but taken from a pool of 10.0.0.0/8. These changes allow all instances
in the same 10.0.0.0/8 to communicate, which should include all VPC
subnetworks.