Allows global access to the token-server LB
Previously, the token-server was only available to instances in the
us-west2 region, which made it impossible for VMs in other regions to
communicate with the token-server to get bootstrap tokens to join the
cluster.

Additionally, add the base VPC network subnet prefix to the networking
variables. We use it to create a firewall rule to allow all subnets to
communicate with each other.