Create MIGs for each virtual "site" (#4)
* Allows all instances in VPC to communicate

The firewall rule was previously incorrect, only allowing instanceds in
the same 10.0.0.0/16 prefix to communicate, which would only let
instances in a single subnet communicate. Our subnets are 10.0.0.0/16,
but taken from a pool of 10.0.0.0/8. These changes allow all instances
in the same 10.0.0.0/8 to communicate, which should include all VPC
subnetworks.

* Simplifies names of a couple TF files

Instead of firewall-rules.tf, firewalls.tf suffices. And removes the
dash from "load-balancers.tf".

* Allows global access to the token-server LB

Previously, the token-server was only available to instances in the
us-west2 region, which made it impossible for VMs in other regions to
communicate with the token-server to get bootstrap tokens to join the
cluster.

Additionally, add the base VPC network subnet prefix to the networking
variables. We use it to create a firewall rule to allow all subnets to
communicate with each other.

* Removes platform instances in favor of MIGs

We intend for virtual machines/nodes to be members of MIGs and to site
behind load balancers. This commit removes the creation of individual
instances, and instead creates MIGs using instance templates, and the
task of creating the VMs is then in the hands of GCP.

* Reorganizes loadbalancers and instancegroups

Previously, loadbalancers.tf contained instance group definitions for
control plane things. Now that we are using instance groups for regular
plaform VMs, it makes more sense for the instancegroups.tf file to only
contain instance group definitions for everything, and for
loadbalancers.tf to only contain the resource releated to the actual
load balancers, namely: static IP address, health checks, backend
services and forwarding rules, which are all part of the LBs.

* Fixes typo plus updates one comment