Modifes resource names to match staging/prod
As long as the names and configuration parameters match for resources
defined in our Terraform configs with those of the actual resources
already in the staging and production, then those resources can be
imported into the Terraform state with no modification or disruption.
This commit attempts to bring the names created by Terraform into line
with how resources are already named in staging and production.