Moves platform-cluster submodule out of mlab-sandbox
Moving the platform-cluster module out of the mlab-sandbox module means
taht will be able to reuse it for mlab-staging and mlab-oti. In those
projects all we have to do is to define new values in a terraform.tfvars
file in those directories.

The awkward symlink of platform-cluster-variables.tf to the variables.tf
file in the platform-cluster module is because each module is insular
and passing input variables to the child module happens in the main.tf
file, in the "module" block. In order to avoid having to keep two
versions of the variables.tf file in sync, we just symlink to the child
module version from the root module. Any changes that need to happen to
the variables.tf file should happen in the platform-cluster module, and
will automatically be picked up by the project root module. Variable
values for the defined variables are set in terraform.tfvars.