Updates boot disk images for regular platform VMs
This should fix the issue of regular platform VMs getting assigned an
InternalIP of the machines IPv6 address instead of its IPv4 address. For
some reason issue is not affecting the API VMs nor the Prometheus VM,
but I'm updating their disk anyway, just so that boot disk versions stay
in sync, and to test tearing everything down, even the control plane,
and putting it back together with no disruption.