Adds local-exec provisioner to instances
GCP supports static, regional, external IPv6 addresses, but the Google
Terraform provider does not yet support this feature. This commit adds a
local-exec provisioner to instances which will remove the ephemeral
external IPv6 from and instance and replace it with a static one. If
this is a new VM and the static address does not already exist, it will
create it. This is a workaround to keep our VM IPv6 configs in sync with
DNS created by siteinfo.