k8s v1.24.15 + new local-exec provisioner for instances (#12)
* Updates boot disk images in sandbox and staging

The new boot disk images container k8s components at v1.24.15.

* Adds local-exec provisioner to instances

GCP supports static, regional, external IPv6 addresses, but the Google
Terraform provider does not yet support this feature. This commit adds a
local-exec provisioner to instances which will remove the ephemeral
external IPv6 from and instance and replace it with a static one. If
this is a new VM and the static address does not already exist, it will
create it. This is a workaround to keep our VM IPv6 configs in sync with
DNS created by siteinfo.

* Adds comment about who to assign static IPv6