Adds small shell wrapper to terraform-apply
Terraform wants to operate on all resources with a fair amount of
concurrency. For many resources this is just fine. However, we don't
want TF to more or less tear down every VM at once, which would cause
downtime and disruption of running experiments. This commit adds a
relatively small shell script to wrap terraform-apply. It iterates over
any updates to instances and applies them one at a time. Once that is
done, it runs terraform-apply on for all other resource types.

This commit also adds a Cloud Build configuration file. It's likely that
the CB service account doesn't have all the necessary permissions to
deploy all the resources we use. Adding the config is a start, at
least.