Adds auto_delete=false to instance boot disks
This setting wasn't needed when Terraform was allowed to do everything
it wanted to do with regard to instances. However, when wrapping TF in a
small shell script to update instances one at a time, TF would delete
the VM and the boot disk would get automatically deleted by GCP. When
TF would go to recreate the instance, it would complain that the
reference boot disk didn't exist, which was true. This should prevent
GCP from auto-deleting the boot disk when a VM gets deleted, which
should prevent this problem.