Adds auto_delete=false to instance boot disks (#14)
* Adds auto_delete=false to instance boot disks

This setting wasn't needed when Terraform was allowed to do everything
it wanted to do with regard to instances. However, when wrapping TF in a
small shell script to update instances one at a time, TF would delete
the VM and the boot disk would get automatically deleted by GCP. When
TF would go to recreate the instance, it would complain that the
reference boot disk didn't exist, which was true. This should prevent
GCP from auto-deleting the boot disk when a VM gets deleted, which
should prevent this problem.

* Removes the secrets.tf file

There is a chicken-and-egg problem here with managing secrets with TF
via Cloud Build. The default Cloud Build service account needs a special
policy binding to be able to create a given secret, and the existing
secrets.tf file is trying to use TF to give the service account that
binding, but of course Cloud Build doesn't have permission to do that,
so an operator needs to manually apply the binding using TF. And you
can't add a policy binding to a secret that doesn't exist, so an
operator needs to manually create the secret too.

Additionally, managing secrets via TF bring up a whole other discussion
of how to keep the secrets secret and not have them stored in the TF
state file in plain text. We can manage secrets with TF, but we will
need to give futher consideration for how privileged we want the default
Cloud Build service account to be, as well as how to store secret data
in a secure fashion.

* Adds comment about why auto_delete=false