Updates build timeout to 10h
I had previously underestimated the time it takes for a VM to be
deleted, recreated, and for ndt-server to start listening on port 443.
The actual time is between 6-7m. This commit updates the build timeout
to reflect how long it will take to update all VMs in the platform
cluster.