Adds new "daemonset" variable
Today, the ndt-virtual DaemonSet's nodeSelectors are "mlab/run=ndt,
mlab/type=virtual", and the mlab/run=ndt label is statically applied to
every platform virtual machine. This commit introduces a new instance
attribute named "daemonset=ndt", which can be overridden in each
individual instance of MIG definition, if we want it to run something
other than the ndt-virtual DaemonSet. The specific motivation for this
feature is the ability to run the ndt-canary DaemonSet on a VM or MIG
instance. I will change the ndt-canary DaemonSet's nodeSelectors to
"mlab/run=ndt-canary, mlab/type=virtual".
.