Adds static IPv6 addresses
When we first started using Terraform, GCP supported static, regional
IPv6 addresses for VMs and loadbalancers, but the Google Terraform
provider did not. However, as of release v4.77.0 of the provider it does
support them. This commit adds static IPv6 resources for platform VMs
and MIG load balancers. It also removes the ugly shell script which was
used to workaround this former limitation.

https://github.com/hashicorp/terraform-provider-google/releases/tag/v4.77.0