Allows overriding machine-type per site
As a general rule, the machine type n2-highcpu-4 seems to be fine for
most regions, but there are some especially busy ones where this isn't
sufficient. For example, in DEL this machine type is overloaded and
there are plenty of OOM events, and CPU usage is probably higher than
we'd like. This commit allows us to selectively change the default
machine type for a VM, and it updates the mlab1-del03 to be
n2-highcpu-8.