Adds static IPv6 addresses (#41)
* Adds static IPv6 addresses

When we first started using Terraform, GCP supported static, regional
IPv6 addresses for VMs and loadbalancers, but the Google Terraform
provider did not. However, as of release v4.77.0 of the provider it does
support them. This commit adds static IPv6 resources for platform VMs
and MIG load balancers. It also removes the ugly shell script which was
used to workaround this former limitation.

https://github.com/hashicorp/terraform-provider-google/releases/tag/v4.77.0

* Allows overriding machine-type per site

As a general rule, the machine type n2-highcpu-4 seems to be fine for
most regions, but there are some especially busy ones where this isn't
sufficient. For example, in DEL this machine type is overloaded and
there are plenty of OOM events, and CPU usage is probably higher than
we'd like. This commit allows us to selectively change the default
machine type for a VM, and it updates the mlab1-del03 to be
n2-highcpu-8.