Creates static internal IPs for VMs
I discovered a bug in our system in which a boot script statically
records the internal IP of the VM in /etc/default/kubelet as the
--node-ip flag to the kubelet. The problem is that the internal
addresses are ephemeral and if the VM needs to be recreated for any
reason, then the value of the --node-ip flag to the kubelet will
differ from the actual internal IP of the machine.

This commit introduces static internal IP addresses for platform nodes
to avoid this problem. The internal address of the machine should never
change, just like the external address.