Removes various variables that didn't change
There were a couple variable defined that didn't ever change between the
various node pools. For simplicity, I removed these variables from the variable
definition and instead use static values in the resource definition. If we ever
decided they need to be different between the node pools, then we can easily
add them back.

Additionally, I was trying to pass a resource reference as a module input
variable for the node pool service accounts, but this is not supported.
Instead, I removed the service_account variable and use a ternary expression to
assign the service account, since only the statitics node pool uses a SA that
is not the default.