Update TF Google provider to v5.12.0 + fix static IPv6 assignment (#59)
* Updates GCP TF provider to v5.12.0

Additionally, comments out the local-exec provisioner for instances which
manually assigns static IPv6s to instances after creation. The hope is that
there was a bug in the previous provider version which caused static IPv6
assignment to not work, and that it is resolved in this newer version.

* Adds field external_ipv6_prefix_length to ipv6_access_config

After some experimentation, I found that adding the prefix length field causes
TF to [re]create a VM with the configured static IPv6 address rather than
giving it an ephemeral one, even though a static one is configured. I don't
understand why this works, but it seems to.

Since static external IPv6 addresses appear to be getting assigned properly
now, this commit also removes the local-exec provisioner that ran the script to
assign static IPv6 addresses properly.

* Updates TF Google provider to v5.12.0

Sandbox was already updated in a previous commit. This updates it for all other
projects.

* Updates boot disk images for VMs in mlab-staging

These images don't contain anything new, but were generated by an epoxy-images
build that had something to do only with physical machine images. However,
updating them here allows us to test VM recreation and static IPv6 assignment
updates that have been made in this repo.