Fully removes multi-VM sites, to be replaced by MIGs
Previously, I had the idea that for virtual "sites" with multiple VMs we would
reduce the virtual site to a single VM and add a MIG. At some point we would
fully retire the standalone VM site. However, this doesn't make sense, since
the MIGs are specifically designed as a drop-in replacement for multi-VM sites.