Start the process of migrating most VMs to MIGs in production (#63)
* Allows overriding machine_type for MIG instances

We allow this for regular, standalone VMs, and this change just allows the same
for MIGs.

* Coverts several standalone VMs to MIGs

https://github.com/m-lab/ops-tracker/issues/1905

* Reduces MIG min replicas from 3 to 1

There is no reason to launch 3 replicas by default if they are unneeded. This
commit makes the minimum number of replicas 1, and will trust that the
auto-scaler will scale out as appropriate to meet demand.

* Adds new subnetwork for region africa-south1

* Fully removes multi-VM sites, to be replaced by MIGs

Previously, I had the idea that for virtual "sites" with multiple VMs we would
reduce the virtual site to a single VM and add a MIG. At some point we would
fully retire the standalone VM site. However, this doesn't make sense, since
the MIGs are specifically designed as a drop-in replacement for multi-VM sites.