Replaces iad08 with iad09 in mlab-oti (#65)
Replaces iad08 with iad09

iad08 is already in use in mlab-staging, and has both an mlab3 and mlab4
machine. This was an early test to see how well it could work to have a
non-mlab4 machine as part of staging. Rather than disturbing that staging
deployment, this change just replaces iad08 in mlab-oti witih iad09.