Adds google.visualizations provider alias to mlab-oti main.tf
We aren't currently using the visualizations module anywhere in the mlab-oti
configuration. However, TF's documentation makes it sound like you should
always list all configuration aliases in the required_providers section, which
is shared between all projects. If you add the alias to the
configuration_aliases list, but don't declare it in the module, TF will emit an
error.