Add new project root module for measurement-lab (#69)
* Adds google.visualizations provider alias to mlab-oti main.tf

We aren't currently using the visualizations module anywhere in the mlab-oti
configuration. However, TF's documentation makes it sound like you should
always list all configuration aliases in the required_providers section, which
is shared between all projects. If you add the alias to the
configuration_aliases list, but don't declare it in the module, TF will emit an
error.

* Adds new project root module for measurement-lab

As a start, it adds the visualizations module, since the public Grafana
instance will run in that project.

* Adds provider alias defs for platform-cluster and visualizations

versions.tf is shared between all projects, and in that file any
configuration_aliases that are listed, must be defined in the root module, even
if those aliases are not used.