Moves CB "thirdparty" config to "non-oti-prod"
cloudbuild-thirdparty.yaml existed only because of required deployments in the
mlab-autojoin project, in which there is no
gcr.io/mlab-autojoin/gcloud-jsonnet-cbif container image, so it manually
pointed to the version of the container image in mlab-oti.

We now have that same need in the measurement-lab project, for which the name
"thirdparty" didn't make sense.

This is an attempt to rename the secondary Cloud Build config file to something
that both the mlab-autojoin and measurement-lab project Cloud Build triggers
can use.

Additionally, I realized that the version terraform that the scripts were
downloading was a bit old. This commit updates the version to the latest, and
additionally makes it an environment variable to avoid so much duplication.