Moves CB "thirdparty" config to "non-oti-prod" (#70)
* Moves CB "thirdparty" config to "non-oti-prod"

cloudbuild-thirdparty.yaml existed only because of required deployments in the
mlab-autojoin project, in which there is no
gcr.io/mlab-autojoin/gcloud-jsonnet-cbif container image, so it manually
pointed to the version of the container image in mlab-oti.

We now have that same need in the measurement-lab project, for which the name
"thirdparty" didn't make sense.

This is an attempt to rename the secondary Cloud Build config file to something
that both the mlab-autojoin and measurement-lab project Cloud Build triggers
can use.

Additionally, I realized that the version terraform that the scripts were
downloading was a bit old. This commit updates the version to the latest, and
additionally makes it an environment variable to avoid so much duplication.

* Escape refs to TF_RELEASE to avoid CB errors

* Removes seconary CB config in favor of CB variable

Rather than having two Cloud Build configurations with the only difference
being the container image path for step #1, this commit removes the secondary
configuration in favor of just creating a user-defined build trigger
substitution in each project pointing to the appropriate container image for
the build.

* Puts an underscore in front of CONTAINER_IMAGE var

* Adds the container image version to the build config

So that the container version is more transparent to a viewer of this
repository, this commit removes the version string from the variable
$_CONTAINER_IMAGE and instead appends it in the build config file. Not only is
this more transparent, but when we need to update the version we just update it
once in this file, rather than having to update variables in build triggers in
every project.