Adds a new "foundations" module (#71)
* Adds a new "common" submodule for generic resources

We currently have submodules specific to certain related infrastructure in GCP
(e.g., platform-cluster, data-pipeline). There are quite a lot of resources in
GCP that do not fall neatly into one piece of infrastructure, or that may be
used by multiple things. This "common" submodule is a place to defined such
resources.

* Adds new common module to mlab-staging and mlab-oti projects

* Renames module "common" to "foundations"

"common" is too generic a word and maybe too broad, and possibly a reserved
sort of word in some contexts.