Adds foundations provider def to two additional projects
Whether a project root module uses a particular submodule or not, it still must
defined the provider alias for all modules, since all project root modules
share a common versions.tf file, which defines which provider aliases should
exist. When I created the foundations module yesterday, I forgot to add the
provider alias to the mlab-autojoin and measurement-lab projects.

There is probably a better way to do this, but I just haven't figured it out
yet. We aren't even using the provider alias features yet, for the most part,
but its nice to have them there in case we need them.