Adds 2nd round of MIGs to production (#77)
https://docs.google.com/document/d/1fX_ndIN2M0MiaguxONTTSXqR2C-JnsSs4PFReY45HuQ/

For this round we will leave the existing standalone VMs in place for now so
that the data from the VMs and the MIGs can be compared side by side for the
same time ranges. And in this way we also don't lose cloud coverage in a
region during the transition. Once we are confident that the data looks good we
can retire the standalone VMs in these regions.