Remove round 2 MIG migration standalone VMs (#81)
* Removes standalone VMs from round 2 of MIG migrations

Round 2 of MIG migrations was successful, and we no longer need the standalone
VMs in those metros. We can now move on to round 3 of MIG migrations.

* Updates disk images of all instance types in sandbox

The images should be the same, but they are built with Packer accessing the
build VM using ssh through Google IAP instead of through normal public ssh
access.

* Updates staging disk images

* Updates staging and production disk images

The images don't contain anything new, but were the result of other changes in
epoxy-images. I update them just so that the running version is consistent with
the latest built image, to keep things easy to understand and for consistency.

* Updates sandbox disk images to latest available