Reduces max instances for MIGs in each project
Previously the max instances for MIGs were 5, 10, 15 for mlab-sandbox,
mlab-staging, mlab-oti, respectively. This PR reduces those to 2, 3, 5, also
respectively. It is good for MIGs to scale out and in, but we need to set a
more reasonable upper bound in order to prevent large traffic spikes from
causing excessive GCP egress network charges.

Additionally, add MIG dfw13 to staging. dfw09 was retired a week ago, but I had
failed to remove it from the TF configs. However, we should have a least one
MIG in each project for testing.