Make load balancers optional for MIGs (#108)
* Updates disk image for platform instances

Updates the write-metadata.sh script to write out the IATA code to
/var/local/metadata

* Makes load balancers optional for MIGs

If M-Lab-managed VMs can use the Autojoin API to get a DNS entry, then they
don't need a load balancer. Not only do load balancers turn out to be
expensive, but they add additional complexity to the system. This commit adds
the necessary changes to make load balancers optional.

* Adds loadbalanced=true to all MIGs in staging and production

This is just to prepopulate the value explicitly, and in case Terraform would
exit with an error if we are trying to check the value of a non-existent field.

* Defines default probability of 1.0 and sets VMs that differ

VMs definitions that don't have a probability set explicitly will get the new
default probability of 1.0.