Add appengine.appViewer permission to autojoin GKE service account (#112)
* Adds port 9100 to public-prometheus-monitoring firewall rule

This exposes node-exporter metrics

* Updates disk images for instance types

* Removes port 9100 from public-prometheus-monitoring firewall rule

I moved the node_exporter port for autonode to 9995, which makes it part of the
existing range of open ports 9990-9999.

* Adds bigquery.dataViewer role to the autojoin GKE service account

We are going to start running an instance of bigquery_exporter in this cluster
and it needs to be able to query the data.

* Removes the bigqueryViewer role from autojoin GKE service account

It turns out that the bigqueryUser role already permits this.

* Gives autojoin GKE service account App Engine Viewer role

gcp-service-discovery needs this role to be able to scrape Prometheus metrics
from AEFlex instances.

* Updates disk images for all projects

The changes shouldn't affect anything in production, but are mostly related to
auto-detecting the primary network interface and disk.